requirements/requirements_board_dev.txt
pytest
Flask==2.0.2
fake-rpi==0.7.1
Mock.GPIO==0.1.8